uniform fairgraph
