+++ bin/fdp-run
@@ -2,13 +2,12 @@
 """FDP run.
 
 Usage:
-  fdp-run --data <data-file>
+  fdp-run
   fdp-run --sparql <sparql-endpoint>
   fdp-run (-h | --help)
 
 Options:
   -h --help     Show this screen.
-  --data <data-file>   ini or ttl file containing FDP data
   --sparql <sparql-endpoint>  sparql endpoint URL
 """
 from docopt import docopt
@@ -16,5 +15,4 @@ from fdp import __version__
 from fdp.fdp import run_app
 
 args = docopt(__doc__, version='FDP run ' + __version__)
-run_app(host='0.0.0.0', port=8080, dataFile=args['--data'],
-        endpoint=args['--sparql'])
+run_app(host='0.0.0.0', port=8080, endpoint=args['--sparql'])
